Added the covergen script to C4 and its corresponding service to svc.
Additionally, two new features have been added to LMSCoverFlow: 1. The visibility of aspects of the card image body is now configurable. Default is 'author', 'title'. 2. The line height of the aforementioned card body aspects is configrable as well. Possible options are Bootstrap line heights or custom line heights which are defined in setGlobalStyles.