Refactor CoverGen to OO version CoverGenerator under Koha/
This refactor was necessary bacause 1) new modules have to use the Koha namespace 2) have to use OO if possible; Started to write unit tests for the CoverGen module; Had to modify the default value of LMSCoverFlow configuration option coverImageExternalSources to be true: More often than not this will be the case for libraries using Koha, just not with LMSCloud. Therefore the default value 'true' makes sense in this case and enables us to explicitly and automatically set it to false if the cover image provider is EKZ. This prevents an unfavorable rewrite of the urlHarvester logic; The api endpoint for generated covers now uses the OO version CoverGenerator under the Koha namespace. The new version has been tested with EKZ, Amazon and Google covers and works as expected while showing improved performance. CoverGenerator is still work in progress, though.