Bug 37141: (follow-up) Check for expired bookings when rendering bookings table
We preserve the existing logic of this bug and add another case
for when we just encounter expired bookings.

In that case we render the fieldset without the 'filtered' class and
determine the value of filter_expired in bookings.js by the class being set.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit e4a6d756763b3aec53d7eea5c1e25e47edb38805)